refactor empty chats, reactions
